Optimize empty String tests on Java 9+
Motivation:

Don't use String#length() to test emptyness. This computes the actual length, which might be non trivial with Java 9 where the underlying encoding is UTF-8 and not US-ASCII.
Use String#isEmpty() instead as it just has to check the length of the nuderlying byte array.

Modifications:

* string.length() == 0 => string.isEmpty()
* string.length() != 0 => !string.isEmpty()
* string.length() > 0 => !string.isEmpty()
* overload StringUtil#isEmpty to accept a String where implementation can be optimized. Original method doesn't seem to be used but it's public so I didn't remove it.

Result:

More performant String emptiness test